use same client instance for both ontology and library daos
